requirements.txt
click==8.0.3
Flask==2.0.2
Flask-MySQLdb==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
python-dotenv==0.19.2
Werkzeug==2.0.2
